Update observed_news_data.csv
The missing column 1 name trips over the linux CRAN checks and I've had a note from CRAN to say it needs fixing.  The vignette removed column one with the default name X1 on windows, but linux version can't seem to find x1.  I've explicitly named it to try to fix.